fix(rs57): rename the X8-side image_pipeline package to x8_image_pipeline
The repo carried two packages named `image_pipeline` — base-side at
base_station/image_pipeline/ and X8-side at firmware/tractor_x8/ —
so whichever side imported first won inside a Python process and the other
failed with ModuleNotFoundError on its own submodules. Tracked as RS-5.7
(slated as IP-W2-10 in run_tests.ps1); cost so far: four test modules that
could not even be collected, one order-dependent deselected test, a CI
exclusion list, per-file process isolation in run_tests.ps1, and a harness
workaround that MERGED both packages into one on-device directory.

The X8-side package is now `x8_image_pipeline` (repo idiom: x8_ prefixes
X8-side things). The base-side package keeps its name, so its larger
consumer surface — image_rx_daemon, web_ui, lora_proto, 17 test files, and
the deployed base station tree — is untouched. image_tx_daemon's import of
base `image_pipeline.frame_format` is likewise unchanged; that cross-package
dependency is why the tractor deploys both packages, and they now coexist
under distinct names instead of merging.

Updated: 4 imports in camera_service.py, 1 in dry_run_w2_05.py, the
path-loading in replay_ipc_capture.py, 6 test modules, 3 harness scripts
(run_live_radio_monitor push + merge comment, run_w2_05_dry_run device dirs,
run_camera_container_x8 presence check), the docker-compose comment, CI, and
current-tree path references in five architecture docs. Dockerfile is
COPY-dot and needed nothing.

Result, verified locally with no exclusions:

  pytest tests -q                1158 passed, 2 skipped, 0 failed
                                 (was 1110 passed with 4 modules ignored
                                  and 1 test deselected)
  test_mode_flip_forces_full_re_encode now passes both standalone AND
  in-suite -- the order-dependence is gone, its assertion untouched.

CI's exclusion list is deleted. run_tests.ps1's per-file isolation is no
longer load-bearing and is kept as a debugging aid, with its header
rewritten to say so.

Bench note: on-device trees under /tmp are rebuilt by the harness at
launch, but any hand-pushed camera stack importing image_pipeline.ipc_to_h747
must be re-pushed after this lands.
