+++ .github/workflows/arduino-ci.yml
@@ -53,28 +53,17 @@ jobs:
           # take out a whole test class. Never use this value on hardware.
           LIFETRAC_FLEET_KEY_HEX: '0102030405060708090a0b0c0d0e0f10'
         # pytest, not `unittest discover`: it collects the same unittest
-        # TestCases PLUS the modules that import pytest directly (~1044 tests
-        # vs 964), and it can express the exclusions below.
+        # TestCases PLUS the modules that import pytest directly.
         #
-        # The four --ignore'd modules and the one --deselect'ed test are ALL
-        # the same unfixed root cause, tracked as RS-5.7: two different
-        # packages are both named `image_pipeline`
-        # (base_station/ and firmware/tractor_x8/), and which one wins is
-        # decided by whichever of the 53 sys.path-mutating test files ran
-        # first. The four cannot even be collected; the deselected test passes
-        # in isolation and fails in-suite, so it is order-dependent, not a
-        # logic defect — do NOT "fix" its assertion.
-        #
-        # These are excluded EXPLICITLY rather than left to fail, so that a red
-        # badge means a real regression. Renaming one package deletes every
-        # line below.
+        # RS-5.7 (resolved 2026-08-16): this step used to carry four
+        # --ignore'd modules and one --deselect, all one root cause — two
+        # packages were both named `image_pipeline` (base_station/ and
+        # firmware/tractor_x8/), and whichever of the sys.path-mutating test
+        # files ran first decided which one won inside a single process. The
+        # X8-side package is now `x8_image_pipeline`, so the full suite
+        # collects and runs with no exclusions.
         run: |
-          python -m pytest tests -q \
-            --ignore=tests/test_data_saving_measures.py \
-            --ignore=tests/test_ipc_to_h747_roundtrip.py \
-            --ignore=tests/test_link_adaptive_budget.py \
-            --ignore=tests/test_x8_tile_encode_cache.py \
-            --deselect "tests/test_x8_encode_mode.py::EncodeCacheInvalidatesOnModeChangeTests::test_mode_flip_forces_full_re_encode"
+          python -m pytest tests -q
 
   firmware-compile-plan:
     runs-on: ubuntu-latest
