+++ LifeTrac-v25/DESIGN-CONTROLLER/MASTER_PLAN.md
@@ -450,7 +450,7 @@ The image-transmission analysis ([`../AI NOTES/2026-04-27_Image_Transmission_InD
 
 The Kurokesu C2 USB camera was bench-validated end-to-end on the tractor X8 on **2026-05-15** using a static `aarch64` FFmpeg binary pushed over ADB: `/dev/video1` ("`C2: C2 KTM-QGFST`") streamed MJPEG @ `1920×1080` and produced a clean 42 KB JPEG (`yuvj420p`, BT.470BG) on first try, after the §8.19 USB host-controller mitigations from the W2-01 work were in place. With pixel acquisition proven, the W2-02 follow-up is now committed as the next active firmware/software workstream — it owns *every* line of code that turns those camera bytes into something the operator's browser actually displays:
 
-**On tractor (Portenta X8 Linux + H747 M7):** the `firmware/tractor_x8/image_pipeline/` capture → tile-diff → ROI-bias → NanoDet detector → tile-delta WebP encoder → fragment → IPC-to-M7 chain (per the [tractor image pipeline checklist in DESIGN-CONTROLLER/TODO.md Phase 5A/5B](TODO.md#phase-5a--foundation-cpu-only-no-ai-accelerator-required-ships-value-alone) and [TRACTOR_NODE.md § Image pipeline](TRACTOR_NODE.md#image-pipeline-portenta-x8-linux-side)), plus the M7 ring-buffer + LoRa TX scheduler that emits `TileDeltaFrame` (topic `0x25`) at P3 without preempting any P0 ControlFrame. Tractor-side detector + `CMD_PERSON_APPEARED` (opcode `0x60`) ride this path.
+**On tractor (Portenta X8 Linux + H747 M7):** the `firmware/tractor_x8/x8_image_pipeline/` capture → tile-diff → ROI-bias → NanoDet detector → tile-delta WebP encoder → fragment → IPC-to-M7 chain (per the [tractor image pipeline checklist in DESIGN-CONTROLLER/TODO.md Phase 5A/5B](TODO.md#phase-5a--foundation-cpu-only-no-ai-accelerator-required-ships-value-alone) and [TRACTOR_NODE.md § Image pipeline](TRACTOR_NODE.md#image-pipeline-portenta-x8-linux-side)), plus the M7 ring-buffer + LoRa TX scheduler that emits `TileDeltaFrame` (topic `0x25`) at P3 without preempting any P0 ControlFrame. Tractor-side detector + `CMD_PERSON_APPEARED` (opcode `0x60`) ride this path.
 
 **On base (Portenta X8 Linux):** the `base_station/image_pipeline/` reassemble → canvas → bg-cache → recolouriser → super-res → independent YOLO safety detector → state-publisher chain (per [BASE_STATION.md § Image pipeline](BASE_STATION.md#image-pipeline-portenta-x8-linux-side)), plus the WebSocket fan-out to the operator browser, the `link_monitor.py` auto-fallback ladder (`CMD_ENCODE_MODE` opcode `0x63`), and the I-frame-loss recovery path (`CMD_REQ_KEYFRAME` opcode `0x62`).
