+++ LifeTrac-v25/DESIGN-CONTROLLER/firmware/tractor_x8/docker-compose.video-test.yml
@@ -6,7 +6,7 @@
 #                         image_frame`. The legacy IpcWriter path falls
 #                         back to /dev/null automatically (no ttymxc3
 #                         device mapping = OSError → /dev/null per
-#                         image_pipeline/ipc_to_h747.py).
+#                         x8_image_pipeline/ipc_to_h747.py).
 #   - tractor-image-tx-v2: same image, runs image_tx_daemon.py. Owns
 #                          /dev/ttymxc3 exclusively, subscribes to the
 #                          MQTT topic above, fragments + HostLink TX_FRAME_REQ
