+++ LifeTrac-v25/DESIGN-CONTROLLER/firmware/tractor_x8/x8_image_pipeline/encode_motion.py
